LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-option-arrow-scroll.optional-expected.txt




FAIL topofdocument: Arrow keys on options should not scroll the document. assert_equals: First option should be focused when opening the select. expected Element node <option>one</option> but got Element node <select id="topofdocument">
  <option>one</option>
  <opt...
FAIL bottomofdocument: Arrow keys on options should not scroll the document. assert_equals: First option should be focused when opening the select. expected Element node <option>one</option> but got Element node <select id="bottomofdocument">
  <option>one</option>
  <...
